+++ sdk/release_notes.md
@@ -4,7 +4,7 @@ card.io Android SDK change log and release notes
 4.0.0
 -----
 * Distribute .aar file instead of .jar and .so files
-* New extras on CardIOActivity
+* New extras on `CardIOActivity`
   - `EXTRA_SCAN_INSTRUCTIONS`
   - `EXTRA_HIDE_CARDIO_LOGO`
   - `EXTRA_SCAN_OVERLAY_LAYOUT_ID`
@@ -12,8 +12,12 @@ card.io Android SDK change log and release notes
   - `EXTRA_RETURN_CARD_IMAGE`
   - `EXTRA_KEEP_APPLICATION_THEME`
   - `EXTRA_USE_PAYPAL_ACTIONBAR_ICON`
+* Remove deprecated extras and methods in `CardIOActivity`
+  - `canReadCardWithCamera(Context)` (use `canReadCardWithCamera()` instead)
+  - `EXTRA_REQUIRE_ZIP` (use `EXTRA_REQUIRE_POSTAL_CODE` instead)
 * New extra `EXTRA_CAPTURED_CARD_IMAGE` returned to calling Activity 
 * New class `BuildConfig`
+
 
 3.2.0
 -----